+++ ios/Classes/flutter_launch_arguments_ffi.c
@@ -1,3 +1,0 @@
-// Relative import to be able to reuse the C sources.
-// See the comment in ../flutter_launch_arguments_ffi.podspec for more information.
-#include "../../src/flutter_launch_arguments_ffi.c"